refactor: extract linear solver classes into separate Array2DLinSolver module
- Move base_linsolver, LU_linsolver, QR_linsolver, CHOL_linsolver, and interface_linsolver class definitions from Array2D.h to Array2DLinSolver.h
- Move corresponding implementations from Array2D.h details namespace to Array2DLinSolver.h
- Remove 538 lines of linear solver code from Array2D.h core implementation
- Reduce Array2D.h coupling by isolating BLAS-dependent linear algebra functionality
