+++ include/Array2D.h
@@ -1617,215 +1617,6 @@ namespace details {
             std::shared_ptr<T_interp> ptr;            
     }; 
     
-    // Linsolver -------------------------------------------------------------//
-    template <typename T_container> 
-    class base_linsolver {    
-        // Base class for linsolver. Note that this class does not take into
-        // account if caller Array2D is const or not.
-        public:      
-            typedef typename T_container::value_type                             value_type;
-            typedef typename T_container::reference                               reference;
-            typedef typename T_container::size_type                               size_type; 
-            typedef typename T_container::difference_type                   difference_type;         
-            typedef typename T_container::coords                                     coords;        
-            typedef typename container_traits<T_container>::nonconst_container    container; 
-            typedef typename container_traits<T_container>::const_container const_container; 
-            
-            friend container;
-            
-            // Rule of 5 and destructor --------------------------------------//
-            base_linsolver() noexcept = default;
-            base_linsolver(const base_linsolver&) = default;
-            base_linsolver(base_linsolver&&) noexcept = default;
-            base_linsolver& operator=(const base_linsolver&) = default;  
-            base_linsolver& operator=(base_linsolver&&) = default;
-            virtual ~base_linsolver() noexcept = default;
-            
-            // Additional Constructors ---------------------------------------//            
-            // Copy input Array2D, and then perform in-place decomposition
-            base_linsolver(const_container &A) : x_buf(A.width(),1), A_factored_ptr(std::make_shared<container>(A)) { }
-            
-            // Conversion ----------------------------------------------------//
-            // This allows tests to see if factorization was successful
-            virtual operator bool() = 0;           
-
-            // Arithmetic methods --------------------------------------------//
-            container& backward_sub(const_container&, const_container&) const;
-            container& forward_sub(const_container&, const_container&) const;
-            virtual const_container& solve(const_container&) const = 0;
-                        
-            // Clone ---------------------------------------------------------//
-            virtual base_linsolver* clone() const = 0; 
-        
-        protected:
-            mutable container x_buf;                    // have a copy per
-            std::shared_ptr<container> A_factored_ptr;  // the factored matrix - immutable
-    };    
-    
-    template <typename T_container> 
-    class LU_linsolver final : public base_linsolver<T_container> { 
-        // Class for the LU solver
-        public:          
-            typedef typename base_linsolver<T_container>::value_type           value_type;
-            typedef typename base_linsolver<T_container>::reference             reference;
-            typedef typename base_linsolver<T_container>::size_type             size_type; 
-            typedef typename base_linsolver<T_container>::difference_type difference_type;         
-            typedef typename base_linsolver<T_container>::coords                   coords;        
-            typedef typename base_linsolver<T_container>::container             container;    
-            typedef typename base_linsolver<T_container>::const_container const_container; 
-                        
-            friend container;
-            
-            // Rule of 5 and destructor --------------------------------------//
-            LU_linsolver() : full_rank() { }
-            LU_linsolver(const LU_linsolver&) = default;
-            LU_linsolver(LU_linsolver&&) noexcept = default;
-            LU_linsolver& operator=(const LU_linsolver&) = default;  
-            LU_linsolver& operator=(LU_linsolver&&) = default;
-            ~LU_linsolver() noexcept = default;
-                        
-            // Additional Constructors ---------------------------------------//            
-            LU_linsolver(const_container&);
-            
-            // Conversion ----------------------------------------------------//
-            // Tests whether system is nonsingular
-            operator bool() override { return full_rank; }
-            
-            // Arithmetic methods --------------------------------------------//
-            const_container& solve(const_container&) const override;
-            
-            // Clone ---------------------------------------------------------//
-            LU_linsolver* clone() const override { return new LU_linsolver(*this); }
-        
-        private:
-            std::shared_ptr<container> piv_ptr; // pivots - immutable
-            bool full_rank;
-    }; 
-    
-    template <typename T_container> 
-    class QR_linsolver final : public base_linsolver<T_container> { 
-        // Class for QR solver
-        public:          
-            typedef typename base_linsolver<T_container>::value_type           value_type;
-            typedef typename base_linsolver<T_container>::reference             reference;
-            typedef typename base_linsolver<T_container>::size_type             size_type; 
-            typedef typename base_linsolver<T_container>::difference_type difference_type;         
-            typedef typename base_linsolver<T_container>::coords                   coords;        
-            typedef typename base_linsolver<T_container>::container             container;  
-            typedef typename base_linsolver<T_container>::const_container const_container; 
-            
-            friend container;
-            
-            // Rule of 5 and destructor --------------------------------------//
-            QR_linsolver() : full_rank(), rank() { }
-            QR_linsolver(const QR_linsolver&) = default;
-            QR_linsolver(QR_linsolver&&) noexcept = default;
-            QR_linsolver& operator=(const QR_linsolver&) = default;  
-            QR_linsolver& operator=(QR_linsolver&&) = default;
-            ~QR_linsolver() noexcept = default;
-                        
-            // Additional Constructors ---------------------------------------//            
-            QR_linsolver(const_container&);
-            
-            // Conversion ----------------------------------------------------//
-            // Tests whether system is nonsingular
-            operator bool() override { return full_rank; }
-                        
-            // Arithmetic methods --------------------------------------------//
-            std::pair<container, value_type> house(const_container&) const;
-            const_container& solve(const_container&) const override;
-            
-            // Clone ---------------------------------------------------------//
-            QR_linsolver* clone() const override { return new QR_linsolver(*this); }
-        
-        private:
-            std::shared_ptr<container> piv_ptr;  // pivots - immutable
-            std::shared_ptr<container> beta_ptr; // beta - immutable
-            bool full_rank;
-            difference_type rank;
-    }; 
-    
-    template <typename T_container> 
-    class CHOL_linsolver final : public base_linsolver<T_container> { 
-        // Class for cholesky solver
-        public:          
-            typedef typename base_linsolver<T_container>::value_type           value_type;
-            typedef typename base_linsolver<T_container>::reference             reference;
-            typedef typename base_linsolver<T_container>::size_type             size_type; 
-            typedef typename base_linsolver<T_container>::difference_type difference_type;         
-            typedef typename base_linsolver<T_container>::coords                   coords;        
-            typedef typename base_linsolver<T_container>::container             container;       
-            typedef typename base_linsolver<T_container>::const_container const_container; 
-            
-            friend container;
-            
-            // Rule of 5 and destructor --------------------------------------//
-            CHOL_linsolver() : pos_def() { }
-            CHOL_linsolver(const CHOL_linsolver&) = default;
-            CHOL_linsolver(CHOL_linsolver&&) noexcept = default;
-            CHOL_linsolver& operator=(const CHOL_linsolver&) = default;  
-            CHOL_linsolver& operator=(CHOL_linsolver&&) = default;
-            ~CHOL_linsolver() noexcept = default;
-                        
-            // Additional Constructors ---------------------------------------//            
-            CHOL_linsolver(const_container &A);
-            
-            // Conversion ----------------------------------------------------//
-            // Tests whether system is positive definite
-            operator bool() override { return pos_def; }
-                        
-            // Solve method --------------------------------------------------//
-            const_container& solve(const_container&) const override;
-            
-            // Clone ---------------------------------------------------------//
-            CHOL_linsolver* clone() const override { return new CHOL_linsolver(*this); }
-        
-        private:
-            bool pos_def;
-    }; 
-
-    template <typename T_linsolver> 
-    class interface_linsolver final { 
-        // This is a container for the base linsolver class to hide the inheritance 
-        // tree. 
-        public:     
-            typedef typename T_linsolver::value_type                 value_type;
-            typedef typename T_linsolver::reference                   reference;
-            typedef typename T_linsolver::size_type                   size_type; 
-            typedef typename T_linsolver::difference_type       difference_type;         
-            typedef typename T_linsolver::coords                         coords;        
-            typedef typename T_linsolver::container                   container;     
-            typedef typename T_linsolver::const_container       const_container; 
-             
-            template <typename T_linsolver2>
-            friend class interface_linsolver;
-            friend container;
-            
-            // Part of rule of 5 ---------------------------------------------//
-            interface_linsolver() noexcept : ptr(nullptr) { }
-            interface_linsolver(const interface_linsolver &linsolver) : ptr(linsolver.ptr ? linsolver.ptr->clone() : nullptr) { }
-            interface_linsolver(interface_linsolver &&linsolver) : ptr(std::move(linsolver.ptr)) { }
-            interface_linsolver& operator=(const interface_linsolver &linsolver) { ptr.reset(linsolver.ptr ? linsolver.ptr->clone() : nullptr); return *this; }
-            interface_linsolver& operator=(interface_linsolver &&linsolver) { ptr = std::move(linsolver.ptr); return *this; }
-            ~interface_linsolver() noexcept = default;
-            
-            // Additional Constructors ---------------------------------------//
-            // explicit is important since ptr is wrapped in shared_ptr, which 
-            // will call delete on ptr when this object goes out of scope.
-            explicit interface_linsolver(T_linsolver *ptr) : ptr(ptr) { }
-            
-            // Conversion ----------------------------------------------------//
-            // Allow implicit conversions to bool to test whether system is "valid"
-            // (i.e. nonsingular for LU decomposition, positive definite for 
-            // cholesky decomposition, etc)
-            operator bool() { return (*ptr); }
-                                                                    
-            // Solve method --------------------------------------------------//
-            const_container& solve(const_container &b) const { return ptr->solve(b); }
-            
-        private:        
-            std::shared_ptr<T_linsolver> ptr;            
-    }; 
 }
 
 // ---------------------------------------------------------------------------//
@@ -3483,323 +3274,7 @@ namespace details {
         }
     } 
         
-    // base_linsolver --------------------------------------------------------//
-    template <typename T_container>    
-    typename base_linsolver<T_container>::container& base_linsolver<T_container>::backward_sub(const_container &A, const_container &b) const {
-        if (A.height() != A.width()) {
-            throw std::invalid_argument("Attempted to perform backward substitution using an Array with size " + A.size_2D_string() +
-                                        ". Array must be square.");
-        }
-        if (A.height() != b.height() || b.width() != 1) {
-            throw std::invalid_argument("Attempted to perform backward substitution using an A with size " + A.size_2D_string() +
-                                        " and a b with size " + b.size_2D_string() + ".");
-        }
-        
-        // Copy b into x_buf, and then solve in-place - since A is square, they
-        // are the same size.
-        std::copy(b.get_pointer(), b.get_pointer() + b.size(), x_buf.get_pointer());
-        for (difference_type p = A.height()-1; p > 0; --p) {
-            if ((std::abs((*A_factored_ptr)(p,p)) < std::numeric_limits<value_type>::epsilon())) {
-                // Set this x value to 1 for singular matrices
-                x_buf(p) = 1;
-            } else {
-                x_buf(p) /= (*A_factored_ptr)(p,p);
-            }
-            
-            for (difference_type p1 = 0; p1 < p; ++p1) {
-                x_buf(p1) -= x_buf(p) * (*A_factored_ptr)(p1,p);
-            }
-        }
-
-        if ((std::abs((*A_factored_ptr)(0,0)) < std::numeric_limits<value_type>::epsilon())) {
-            x_buf(0) = 1;
-        } else {
-            x_buf(0) /= (*A_factored_ptr)(0,0);            
-        }
-
-        return x_buf;
-    }
-    
-    template <typename T_container>    
-    typename base_linsolver<T_container>::container& base_linsolver<T_container>::forward_sub(const_container &A, const_container &b) const {
-        if (A.height() != A.width()) {
-            throw std::invalid_argument("Attempted to perform forward substitution using an Array with size " + A.size_2D_string() +
-                                        ". Array must be square.");
-        }
-        if (A.height() != b.height() || b.width() != 1) {
-            throw std::invalid_argument("Attempted to perform forward substitution using an A with size " + A.size_2D_string() +
-                                        " and a b with size " + b.size_2D_string() + ".");
-        }
-           
-        // Copy b into x_buf, and then solve in-place - since A is square, they
-        // are the same size.
-        std::copy(b.get_pointer(), b.get_pointer() + b.size(), x_buf.get_pointer());
-        for (difference_type p = 0; p < A.height() - 1; ++p) {
-            if ((std::abs((*A_factored_ptr)(p,p)) < std::numeric_limits<value_type>::epsilon())) {
-                // Set this x value to 1 for singular matrices
-                x_buf(p) = 1;
-            } else {
-                x_buf(p) /= (*A_factored_ptr)(p,p);
-            }
-            
-            for (difference_type p1 = p + 1; p1 < A_factored_ptr->height(); ++p1) {
-                x_buf(p1) -= x_buf(p) * (*A_factored_ptr)(p1,p);
-            }
-        }
-
-        if ((std::abs((*A_factored_ptr)(last,last)) < std::numeric_limits<value_type>::epsilon())) {
-            x_buf(last) = 1;
-        } else {
-            x_buf(last) /= (*A_factored_ptr)(last,last);            
-        }
-
-        return x_buf;
-    }
-    
-    // LU_linsolver ----------------------------------------------------------//
-    template <typename T_container>    
-    LU_linsolver<T_container>::LU_linsolver(const_container &A) : base_linsolver<container>(A), piv_ptr(std::make_shared<container>(A.height(),1)), full_rank(true) {
-        if (A.height() != A.width()) {
-            throw std::invalid_argument("Attempted to perform LU decomposition on array of size " + A.size_2D_string() + 
-                                        ". Array must be square.");  
-        }
-                
-        // From Matrix Operations Golab & Van Loan 
-        // LU decomposition with partial pivoting
-        for (difference_type p = 0; p < this->A_factored_ptr->height()-1; ++p) {
-            // Find mu, the pivot
-            difference_type mu = std::max_element(&(*this->A_factored_ptr)(p,p), &(*this->A_factored_ptr)(0,p) + this->A_factored_ptr->height()) - &(*this->A_factored_ptr)(0,p);
-            (*this->piv_ptr)(p) = mu; // store it
-            
-            // Swap rows
-            for (difference_type p2 = p; p2 < this->A_factored_ptr->width(); ++p2) {
-                value_type buf = (*this->A_factored_ptr)(p,p2);
-                (*this->A_factored_ptr)(p,p2) = (*this->A_factored_ptr)(mu,p2);
-                (*this->A_factored_ptr)(mu,p2) = buf;
-            }
-                        
-            // Test to make sure index at (i,i) isn't close to zero
-            if (std::abs((*this->A_factored_ptr)(p,p)) < std::numeric_limits<value_type>::epsilon()) {
-                full_rank = false;
-            } else {
-                for (difference_type p1 = p+1; p1 < this->A_factored_ptr->height(); ++p1) {
-                    (*this->A_factored_ptr)(p1,p) /= (*this->A_factored_ptr)(p,p); 
-                }
-                              
-                for (difference_type p2 = p+1; p2 < this->A_factored_ptr->width(); ++p2) {
-                    for (difference_type p1 = p+1; p1 < this->A_factored_ptr->height(); ++p1) {
-                        (*this->A_factored_ptr)(p1,p2) -= (*this->A_factored_ptr)(p1,p) * (*this->A_factored_ptr)(p,p2);
-                    }
-                }
-            }
-        }
-        
-        // Test last diagonal since loop exits before then
-        if (std::abs((*this->A_factored_ptr)(last,last)) < std::numeric_limits<value_type>::epsilon()) {
-            full_rank = false;
-        }
-    }      
-        
-    template <typename T_container>    
-    typename LU_linsolver<T_container>::const_container& LU_linsolver<T_container>::solve(const_container &b) const {
-        if ((*this->A_factored_ptr).height() != b.height() || b.width() != 1) {
-            throw std::invalid_argument("Attempted to solve LU decomposition using b of size " + b.size_2D_string() + 
-                                        " on an Array of size " + (*this->A_factored_ptr).size_2D_string() + ".");  
-        }
-        
-        // For singular matrices, this will substitute 1 for indeterminant coordinates
-        // t(P)LUx = b 
-        // 1) Solve Ly = Pb
-        // 2) Solve Ux = y
-        
-        // copy b into x_buf first
-        std::copy(b.get_pointer(), b.get_pointer() + b.size(), this->x_buf.get_pointer());
-        for (difference_type p = 0; p < b.height()-1; ++p) {
-            // Swap element
-            value_type buf = this->x_buf((*piv_ptr)(p));
-            this->x_buf((*piv_ptr)(p)) = this->x_buf(p);
-            this->x_buf(p) = buf;
-
-            // Scale
-            for (difference_type p1 = p+1; p1 < b.height(); ++p1) {
-                this->x_buf(p1) -= this->x_buf(p) * (*this->A_factored_ptr)(p1,p);
-            }
-        }        
-        // Solve Ux = y and return it
-        return this->backward_sub((*this->A_factored_ptr), this->x_buf);
-    }
-    
-    // QR_linsolver ----------------------------------------------------------//
-    template <typename T_container>    
-    QR_linsolver<T_container>::QR_linsolver(const_container &A) : base_linsolver<container>(A), piv_ptr(std::make_shared<container>(A.width(),1)), beta_ptr(std::make_shared<container>(A.width(),1)), full_rank(true), rank(0) {        
-        // From Matrix Operations Golab & Van Loan 
-        // QR decomposition with column pivoting
-        
-        // QR related algorithms use the Array2D interface since most of the 
-        // bottleneck should be computational. This simplifies some arithmetic.
-        
-        container c(A.width(),1);
-        for (difference_type p2 = 0; p2 < A.width(); ++p2) {
-            c(p2) = dot(A(all,p2),A(all,p2));
-        }    
-        value_type tau = max(c);
-        difference_type k = find(c == tau);
-        difference_type min_dim = std::min(A.height(), A.width());
-        while (std::abs(tau) >= std::numeric_limits<value_type>::epsilon() && tau > 0) {
-            ++rank;
-            (*piv_ptr)(rank-1) = k;
-
-            // Swap columns
-            container col_buf = (*this->A_factored_ptr)(all,k);
-            (*this->A_factored_ptr)(all,k) = (*this->A_factored_ptr)(all,rank-1);
-            (*this->A_factored_ptr)(all,rank-1) = col_buf;
-            
-            // Swap c
-            value_type buf_c = c(k);
-            c(k) = c(rank-1);
-            c(rank-1) = buf_c;
-            
-            // Get householder vector
-            auto h_pair = house((*this->A_factored_ptr)({rank-1,last},rank-1));
-            (*beta_ptr)(rank-1) = h_pair.second; // store beta
-                        
-            (*this->A_factored_ptr)({rank-1,last},{rank-1,last}) = (*this->A_factored_ptr)({rank-1,last},{rank-1,last}) - h_pair.second * h_pair.first * ( t(h_pair.first) * (*this->A_factored_ptr)({rank-1,last},{rank-1,last}) );
-            (*this->A_factored_ptr)({rank,last},rank-1) = h_pair.first({1,(*this->A_factored_ptr).height()-rank});
-            
-            for (difference_type p2 = rank; p2 < (*this->A_factored_ptr).width(); ++p2) {
-                c(p2) = c(p2) - std::pow((*this->A_factored_ptr)(rank-1,p2),2);
-            }
-                        
-            if (rank < min_dim) {
-                tau = max(c({rank,last}));
-                k = find(c == tau, rank);
-            } else {
-                tau = 0;
-            }
-        }        
-        // Test if full (column) rank
-        full_rank = (rank == (*this->A_factored_ptr).width());
-    }          
-    
-    template <typename T_container>    
-    std::pair<typename QR_linsolver<T_container>::container,typename QR_linsolver<T_container>::value_type> QR_linsolver<T_container>::house(const_container &x) const {
-        if (x.width() != 1) {
-            throw std::invalid_argument("Attempted to get household vector for vector of size " + x.size_2D_string() + 
-                                        ". Vector must be a column vector.");  
-        }
-        
-        value_type sigma = dot(x({1,last}),x({1,last}));
-        container v(x.height(),1,1);
-        v({1,last}) = x({1,last});
-        value_type beta = 0;
-        if (std::abs(sigma) > std::numeric_limits<value_type>::epsilon()) {
-            value_type mu = std::sqrt(std::pow(x(0),2) + sigma);
-            if (std::abs(x(0)) < std::numeric_limits<value_type>::epsilon() || x(0) < 0) {
-                v(0) = x(0) - mu;
-            } else {
-                v(0) = -sigma / (x(0) + mu);
-            }
-            beta = 2 * std::pow(v(0),2) / (sigma + std::pow(v(0),2));
-            v = v/v(0);
-        }
-        
-        return {std::move(v),std::move(beta)};
-    }
-    
-    template <typename T_container>    
-    typename QR_linsolver<T_container>::const_container& QR_linsolver<T_container>::solve(const_container &b) const {
-        if ((*this->A_factored_ptr).height() != b.height() || b.width() != 1) {
-            throw std::invalid_argument("Attempted to solve QR decomposition using b of size " + b.size_2D_string() + 
-                                        " on an Array of size " + (*this->A_factored_ptr).size_2D_string() + ".");  
-        }
-        
-        // Obtains the least squares solution for over determined systems. For
-        // non full-rank systems, this will determine the basic solution.
-        // QRt(P)x = b
-        // 1) Find Rt(P)x = t(Q)b = y
-        // 2) Solve Rt(P)x = y
-        container y(b);
-        container v(b.height(),1); // buffer for householder vector
-        for (difference_type p = 0; p < rank; ++p) {
-            v(p) = 1;
-            v({p+1,last}) = (*this->A_factored_ptr)({p+1,last},p);
-            y({p,last}) = y({p,last}) - (*beta_ptr)(p) * v({p,last}) * (t( v({p,last}) ) * y({p,last}));
-        }
-                        
-        // Solve for t(P)*x - stores results in x_buf
-        this->backward_sub((*this->A_factored_ptr)({0,rank-1},{0,rank-1}),y({0,rank-1}));
-        
-        // Solve for basic solution
-        this->x_buf({rank,last}) = 0; // set rest of x is set to zero for basic solution
-        for (difference_type p = rank; p > 0; --p) {
-            value_type buf = this->x_buf((*piv_ptr)(p-1));
-            this->x_buf((*piv_ptr)(p-1)) = this->x_buf(p-1);
-            this->x_buf(p-1) = buf;    
-        }
-        
-        return this->x_buf;
-    }
-    
-    // CHOL_linsolver ----------------------------------------------------------//
-    template <typename T_container>    
-    CHOL_linsolver<T_container>::CHOL_linsolver(const_container &A) : base_linsolver<container>(A), pos_def(true) {
-        if (A.height() != A.width()) {
-            throw std::invalid_argument("Attempted to perform Cholesky decomposition on array of size " + A.size_2D_string() + 
-                                        ". Array must be square.");  
-        }
-        
-        // From Matrix Operations Golab & Van Loan 
-        // Returns Cholesky decomposition in the lower AND upper half of the matrix.
-        // This allows forward and backward substitution to be used without transposing.
-        for (difference_type p = 0; p < (*this->A_factored_ptr).height(); ++p) {
-            if (p > 0) {
-                for (difference_type p1 = this->A_factored_ptr->height() - 1; p1 >= p; --p1) {
-                    double buf = 0.0;
-                    for (difference_type p2 = 0; p2 < p; ++p2) {
-                        buf += (*this->A_factored_ptr)(p1,p2) * (*this->A_factored_ptr)(p,p2);
-                    }
-                    (*this->A_factored_ptr)(p1,p) -= buf;
-                }
-            }
-
-            // Check to make sure diagonal element is greater than zero
-            if ((*this->A_factored_ptr)(p,p) >= std::numeric_limits<value_type>::epsilon()) {
-                double diag_sqrt = std::sqrt((*this->A_factored_ptr)(p,p));
-                for (difference_type p1 = p; p1 < this->A_factored_ptr->height(); ++p1) {
-                    (*this->A_factored_ptr)(p1,p) /= diag_sqrt;
-                } 
-            } else {
-                pos_def = false;
-                return; // Must return because Cholesky decomposition does not exist
-            }
-        } 
-        
-        // Copy lower half into upperhalf
-        for (difference_type p2 = 0; p2 < this->A_factored_ptr->width() - 1; ++p2) {
-            for (difference_type p1 = p2 + 1; p1 < this->A_factored_ptr->height(); ++p1) {
-                (*this->A_factored_ptr)(p2,p1) = (*this->A_factored_ptr)(p1,p2);
-            }
-        }
-    }      
-        
-    template <typename T_container>    
-    typename CHOL_linsolver<T_container>::const_container& CHOL_linsolver<T_container>::solve(const_container &b) const {
-        if (!pos_def) {
-            throw std::invalid_argument("Attempted to solve Cholesky decomposition with a non positive definite matrix.");  
-        }
-        if ((*this->A_factored_ptr).height() != b.height() || b.width() != 1) {
-            throw std::invalid_argument("Attempted to solve Cholesky decomposition using b of size " + b.size_2D_string() + 
-                                        " on an Array of size " + (*this->A_factored_ptr).size_2D_string() + ".");  
-        }
-        
-        // LUx = b 
-        // 1) Solve Ly = b
-        // 2) Solve Ux = y
-        // Since Cholesky decomp stores factorization in lower and upper half,
-        // you can use backward_sub without transposing the matrix.
-        return this->backward_sub((*this->A_factored_ptr), this->forward_sub((*this->A_factored_ptr), b));
-    }
-}
+} // namespace details
 
 // General Functions ---------------------------------------------------------//
 template <typename T = double,typename T_alloc = std::allocator<T>>
@@ -3818,7 +3293,7 @@ Array2D<T,T_alloc> eye(typename Array2D<T,T_alloc>::difference_type n, T = T(), 
     return A;
 }
 
-}
+} // namespace ncorr
 
 #include "Array2DLinSolver.h"
